test: share google media picker setup
